Develop

+++ db/migrate/20160304140623_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration
   def change
     create_table :products do |t|
     	t.integer :product_id, null: false
-    	t.float :product_price, null: false
+    	t.integer :product_price, null: false
     	t.integer :cart_id
     	t.string :product_name, null: false
     	t.string :product_description, null: false
